+++ src/tsdl_stubs.c
@@ -5,7 +5,21 @@
 
 /* This is just here for ocamlbuild to generate a correct dlltsdl.so object */
 
+#ifdef _MSC_VER
+/* MSVC requires at least one extern function from SDL2 to be used.
+   If not, MSVC will not link the SDL2 import library and subsequently the
+   SDL2 DLL will not be implicitly linked (aka. load-time dynamic linking)
+   to the TSDL stub DLL. That can be verified with Unix `ldd` or MSVC
+   `dumpbin /imports`.
+   ctypes-foreign, in particular ctypes_win32_dlsym_rtld_default(), requires
+   that all foreign DLLs are already mapped into the process address space.
+   Implicit linking is the simplest way to do that.
+ */
+#include "SDL.h"
+void tsdl_nop (void) { SDL_WasInit(0); return; }
+#else
 void tsdl_nop (void) { return; }
+#endif
 
 /*---------------------------------------------------------------------------
    Copyright (c) 2013 The tsdl programmers